Add files via upload
Question 7: What can we infer from the datasets about the history/economy/world back- ground? The Soviet Union/Russia and the United States are the two major countries in aerospace exploration and missions, while more and more countries decided to start their missions. What can we reflect from this change?

The inferences that can be made from this line plot can be formed into a timeseries of the events that caused these fluctuations.

Cold War & Shuttle era (1960s - 1970s): There was a huge U.S. spike in the mid 1980s, which was the escalation of astronaut classes to staff more frequent mission, called the Shuttle era.

USSR & Russia split: The USSR line plateaus to zero by the 1990s and Russia rises, which is consistent with 1991 transition caused by the fall of the Soviet Union. We can see Russia's yearly count stay in the low-moderate range which matches their steady number of missions.

Late 1990s and 2000s ISS cooperation: The late 1990s and 2000s shows consistent U.S. activity and Russian trickles, this was during the Shuttle-Mir, an ISS assembly and operation. 

Rise of "Others" category: The "others" category is near-zero for a while, and only rises after the 1990s, which portrays the internationalzation of the ISS, where other states like Japan, Canada etc. begin structured flight opportunities and get shared seating in the ISS.

Policy & Budget Changes in the early 2010s for the US: There is a dip in 2011-2016 era for th eUS which matches the retirement of shuttles. On the other hand, Russia starts carrying more clew deliveries via Soyuz. However, later on the US rises again with new classes and commercial crew.

Inferences: When there's a stable platform and funding (Shuttles, ISS, and Commercial Crews), hte US has large selection waves. But when costs and vehicles change, the count dips. Russia has kept a small, but steady flow, and more countries are appearing more and more once costs and risks are shared (as well as partnerships).